+++ fmriprep/info.py
@@ -5,7 +5,7 @@
 Base module variables
 """
 
-__version__ = '0.6.1-dev'
+__version__ = '1.0.0-rc3-dev'
 __author__ = 'The CRN developers'
 __copyright__ = 'Copyright 2017, Center for Reproducible Neuroscience, Stanford University'
 __credits__ = ['Craig Moodie', 'Ross Blair', 'Oscar Esteban', 'Chris Gorgolewski',
@@ -65,7 +65,7 @@ REQUIRES = [
     'nibabel>=2.1.0',
     'pandas',
     'grabbit',
-    'pybids>=0.2.1',
+    'pybids>=0.3',
     'nitime',
     'niworkflows>=0.1.6',
     'statsmodels',