+++ fmriprep/interfaces/__init__.py
@@ -5,12 +5,14 @@
 from .bids import (
     ReadSidecarJSON, DerivativesDataSink, BIDSDataGrabber, BIDSFreeSurferDir, BIDSInfo
 )
-from .images import IntraModalMerge, InvertT1w, ValidateImage, ConformSeries
+from .images import (
+    IntraModalMerge, InvertT1w, ValidateImage, TemplateDimensions, Conform, Reorient
+)
 from .freesurfer import (
     StructuralReference, MakeMidthickness, FSInjectBrainExtracted, FSDetectInputs
 )
 from .surf import NormalizeSurf, GiftiNameSource, GiftiSetAnatomicalStructure
 from .reports import SubjectSummary, FunctionalSummary, AboutSummary
-from .utils import ApplyMask, TPM2ROI, ConcatROIs, CombineROIs, AddTSVHeader
+from .utils import TPM2ROI, ConcatROIs, CombineROIs, AddTSVHeader
 from .fmap import FieldEnhance
 from .confounds import GatherConfounds, ICAConfounds